+++ migrations/token/ERC20/polygon_bridging/pos-portal/ERC20EscrowPredicate_deploy.js
@@ -1,0 +1,10 @@
+const Contract_deploy = require('../../../../../src/templates/Contract/deploy');
+const {getNamedAccount} = require('../../../../../src/helpers/templates');
+const {skipIfChainTypeIsNot} = require('../../../../../src/helpers/common');
+
+module.exports = Contract_deploy('ERC20EscrowPredicate', {
+  contract: 'ERC20EscrowPredicate',
+  args: [{name: 'RootChainManager', value: getNamedAccount('POSPortal_RootChainManager_Proxy')}],
+});
+module.exports.skip = skipIfChainTypeIsNot('ethereum');
+module.exports.tags = ['ERC20', 'PolygonPOSPortal'];